Added safety net for grabbing news if not connected to internet
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@1496 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/view/swing/start/StartPanel.java
@@ -391,13 +391,11 @@ public class StartPanel extends javax.swing.JPanel implements ComponentListener 
             File newsFile = DownloadFile.downloadFile(new URL(BrowserSettings.url_news), DirectorySettings.getTmpDirectory());
             if (newsFile != null) {
                 p = parseNewsFile(newsFile);
-            } else {
-                p.add(this.createLabel("Problem getting news"));
             }
-            newsFile.delete();
+            if (newsFile.exists()) { newsFile.delete(); }
 
         } catch (Exception ex) {
-            ex.printStackTrace();
+            p.add(this.createLabel("Problem getting news"));
         }
 
         return p;
